Remove webpacks and updated package.json
