+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "test": "karma start",
     "clean": "rimraf dist",
-    "build:webpack": "webpack --config webpack.config.prod.js --history-api-fallback",
+    "build:webpack": "webpack --config webpack.config.js --history-api-fallback",
     "build:cp": "cp index.html dist/index.html",
     "build": "npm run clean && npm run build:webpack && npm run build:cp",
     "start": "node devServer.js",
